+++ src/command_modules/azure-cli-container/azure/cli/command_modules/container/_help.py
@@ -41,6 +41,8 @@ helps['container create'] = """
           text: az container create -g MyResourceGroup --name myapp --log-analytics-workspace myworkspace
         - name: Create a container group with a system assigned identity.
           text: az container create -g MyResourceGroup --name myapp --image myimage:latest --assign-identity
+        - name: Create a container group with a system assigned identity. The group will have a 'Contributor' role with access to a storage account.
+          text: az container create -g MyResourceGroup --name myapp --image myimage:latest --assign-identity --scope /subscriptions/99999999-1bf0-4dda-aec3-cb9272f09590/MyResourceGroup/myRG/providers/Microsoft.Storage/storageAccounts/storage1
         - name: Create a container group with a user assigned identity.
           text: az container create -g MyResourceGroup --name myapp --image myimage:latest --assign-identity  /subscriptions/mySubscrpitionId/resourcegroups/myRG/providers/Microsoft.ManagedIdentity/userAssignedIdentities/myID
         - name: Create a container group with both system and user assigned identity.