+++ src/command_modules/azure-cli-container/azure/cli/command_modules/container/_params.py
@@ -11,7 +11,7 @@ from azure.cli.core.commands.parameters import (get_enum_type,
 from azure.cli.core.commands.validators import get_default_location_from_resource_group
 from azure.mgmt.containerinstance.models import (
     ContainerGroupRestartPolicy, OperatingSystemTypes, ContainerNetworkProtocol)
-from ._validators import (validate_volume_mount_path, validate_secrets, validate_subnet,
+from ._validators import (validate_volume_mount_path, validate_secrets, validate_subnet, validate_msi,
                           validate_gitrepo_directory, validate_network_profile, validate_image)
 
 # pylint: disable=line-too-long
@@ -78,7 +78,12 @@ def load_arguments(self, _):
         c.argument('secrets', secrets_type)
         c.argument('secrets_mount_path', validator=validate_volume_mount_path, help="The path within the container where the secrets volume should be mounted. Must not contain colon ':'.")
         c.argument('file', options_list=['--file', '-f'], help="The path to the input file.")
-        c.argument('assign_identity', nargs='*', arg_group='Managed Service Identity', help="Space-separated list of assigned identities. Assigned identities are either user assigned identities (resource IDs) and / or the system assigned identity ('[system]'). See examples for more info.")
+
+    with self.argument_context('container create', arg_group='Managed Service Identity') as c:
+        c.argument('assign_identity', nargs='*', validator=validate_msi, help="Space-separated list of assigned identities. Assigned identities are either user assigned identities (resource IDs) and / or the system assigned identity ('[system]'). See examples for more info.")
+        c.argument('identity_scope', options_list=['--scope'], help="Scope that the system assigned identity can access")
+        c.argument('identity_role', options_list=['--role'], help="Role name or id the system assigned identity will have")
+        c.ignore('identity_role_id')
 
     with self.argument_context('container create', arg_group='Network') as c:
         c.argument('network_profile', network_profile_type)